+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe CLI package will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Ensure `zowe config auto-init` command saves the `rejectUnauthorized` value. [#1109](https://github.com/zowe/zowe-cli/issues/1109)
+
 ## `7.0.0-next.202201111811`
 
 - BugFix: Update Imperative to absorb bugfixes introduced in version `5.0.0-next.202201102100`.